+++ src/main/java/orgs/tuasl_clint/models2/FactoriesSQLite/ChatParticipantFactory.java
@@ -4,7 +4,7 @@ import orgs.tuasl_clint.models2.ChatParticipant;
 import orgs.tuasl_clint.utils.DatabaseConnectionSQLite;
 
 import java.sql.*;
-import java.math.BigInteger;
+
 import java.util.ArrayList;
 import java.util.List;
 
@@ -15,22 +15,22 @@ public class ChatParticipantFactory {
 
     public ChatParticipant createFromResultSet(ResultSet rs) throws SQLException {
         return new ChatParticipant(
-                BigInteger.valueOf(rs.getLong("chat_participant_id")),
-                BigInteger.valueOf(rs.getLong("chat_id")),
-                BigInteger.valueOf(rs.getLong("user_id")),
+                (rs.getLong("chat_participant_id")),
+                (rs.getLong("chat_id")),
+                (rs.getLong("user_id")),
                 ChatParticipant.ChatParticipantRole.fromString(rs.getString("role")),
                 rs.getTimestamp("joined_at"),
                 rs.getTimestamp("muted_until"),
                 rs.getInt("is_pinned") == 1,
                 rs.getInt("unread_count"),
-                rs.getLong("last_read_message_id") != 0 ? BigInteger.valueOf(rs.getLong("last_read_message_id")) : null
+                rs.getLong("last_read_message_id") != 0 ? (rs.getLong("last_read_message_id")) : null
         );
     }
 
-    public ChatParticipant findById(BigInteger chatParticipantId) throws SQLException {
+    public ChatParticipant findById(long chatParticipantId) throws SQLException {
         String sql = "SELECT * FROM chat_participants WHERE chat_participant_id = ?";
         try (PreparedStatement statement = DatabaseConnectionSQLite.getInstance().getConnection().prepareStatement(sql)) {
-            statement.setLong(1, chatParticipantId.longValue());
+            statement.setLong(1, chatParticipantId);
             try (ResultSet rs = statement.executeQuery()) {
                 if (rs.next()) {
                     return createFromResultSet(rs);
@@ -40,11 +40,11 @@ public class ChatParticipantFactory {
         return null;
     }
 
-    public List<ChatParticipant> findByChatId(BigInteger chatId) throws SQLException {
+    public List<ChatParticipant> findByChatId(long chatId) throws SQLException {
         String sql = "SELECT * FROM chat_participants WHERE chat_id = ?";
         List<ChatParticipant> participants = new ArrayList<>();
         try (PreparedStatement statement = DatabaseConnectionSQLite.getInstance().getConnection().prepareStatement(sql)) {
-            statement.setLong(1, chatId.longValue());
+            statement.setLong(1, chatId);
             try (ResultSet rs = statement.executeQuery()) {
                 while (rs.next()) {
                     participants.add(createFromResultSet(rs));
@@ -54,11 +54,11 @@ public class ChatParticipantFactory {
         return participants;
     }
 
-    public List<ChatParticipant> findByUserId(BigInteger userId) throws SQLException {
+    public List<ChatParticipant> findByUserId(long userId) throws SQLException {
         String sql = "SELECT * FROM chat_participants WHERE user_id = ?";
         List<ChatParticipant> participants = new ArrayList<>();
         try (PreparedStatement statement = DatabaseConnectionSQLite.getInstance().getConnection().prepareStatement(sql)) {
-            statement.setLong(1, userId.longValue());
+            statement.setLong(1, userId);
             try (ResultSet rs = statement.executeQuery()) {
                 while (rs.next()) {
                     participants.add(createFromResultSet(rs));
@@ -68,11 +68,11 @@ public class ChatParticipantFactory {
         return participants;
     }
 
-    public ChatParticipant findByChatAndUser(BigInteger chatId, BigInteger userId) throws SQLException {
+    public ChatParticipant findByChatAndUser(long chatId, long userId) throws SQLException {
         String sql = "SELECT * FROM chat_participants WHERE chat_id = ? AND user_id = ?";
         try (PreparedStatement statement = DatabaseConnectionSQLite.getInstance().getConnection().prepareStatement(sql)) {
-            statement.setLong(1, chatId.longValue());
-            statement.setLong(2, userId.longValue());
+            statement.setLong(1, chatId);
+            statement.setLong(2, userId);
             try (ResultSet rs = statement.executeQuery()) {
                 if (rs.next()) {
                     return createFromResultSet(rs);
