+++ src/main/java/orgs/tuasl_clint/models2/FactoriesSQLite/MediaFactory.java
@@ -4,7 +4,7 @@ import orgs.tuasl_clint.models2.Media;
 import orgs.tuasl_clint.utils.DatabaseConnectionSQLite;
 
 import java.sql.*;
-import java.math.BigInteger;
+
 import java.util.ArrayList;
 import java.util.List;
 
@@ -15,12 +15,12 @@ public class MediaFactory {
 
     public Media createFromResultSet(ResultSet rs) throws SQLException {
         return new Media(
-                BigInteger.valueOf(rs.getLong("media_id")),
-                rs.getLong("uploader_user_id") != 0 ? BigInteger.valueOf(rs.getLong("uploader_user_id")) : null,
+                (rs.getLong("media_id")),
+                rs.getLong("uploader_user_id") != 0 ? (rs.getLong("uploader_user_id")) : null,
                 rs.getString("file_name"),
                 rs.getString("file_path_or_url"),
                 rs.getString("mime_type"),
-                BigInteger.valueOf(rs.getLong("file_size_bytes")),
+                (rs.getLong("file_size_bytes")),
                 rs.getString("thumbnail_url"),
                 rs.getObject("duration_seconds") != null ? rs.getInt("duration_seconds") : null,
                 rs.getObject("width") != null ? rs.getInt("width") : null,
@@ -29,10 +29,10 @@ public class MediaFactory {
         );
     }
 
-    public Media findById(BigInteger mediaId) throws SQLException {
+    public Media findById(long mediaId) throws SQLException {
         String sql = "SELECT * FROM media WHERE media_id = ?";
         try (PreparedStatement statement = DatabaseConnectionSQLite.getInstance().getConnection().prepareStatement(sql)) {
-            statement.setLong(1, mediaId.longValue());
+            statement.setLong(1, mediaId);
             try (ResultSet rs = statement.executeQuery()) {
                 if (rs.next()) {
                     return createFromResultSet(rs);
@@ -42,11 +42,11 @@ public class MediaFactory {
         return null;
     }
 
-    public List<Media> findByUploader(BigInteger uploaderUserId) throws SQLException {
+    public List<Media> findByUploader(long uploaderUserId) throws SQLException {
         String sql = "SELECT * FROM media WHERE uploader_user_id = ?";
         List<Media> mediaList = new ArrayList<>();
         try (PreparedStatement statement = DatabaseConnectionSQLite.getInstance().getConnection().prepareStatement(sql)) {
-            statement.setLong(1, uploaderUserId.longValue());
+            statement.setLong(1, uploaderUserId);
             try (ResultSet rs = statement.executeQuery()) {
                 while (rs.next()) {
                     mediaList.add(createFromResultSet(rs));
